+++ R/glimpse-Spat.R
@@ -1,15 +1,15 @@
-#' Get a glimpse of your Spat* objects
+#' Get a nice glimpse of your `Spat*` objects
 #'
 #' @description
 #' `glimpse()` is like a transposed version of [print()]: layers/columns run
 #' down the  page, and data runs across. This makes it possible to see every
-#' layer/column in a Spat* object.
+#' layer/column in a `Spat*` object.
 #'
 #' @export
 #' @rdname glimpse.Spat
 #' @name glimpse.Spat
 #'
-#' @seealso [dplyr::glimpse()]
+#' @seealso [tibble::print.tbl_df()]
 #'
 #' @family dplyr.cols
 #' @family dplyr.methods
@@ -21,18 +21,22 @@
 #' be used within a data pipeline.
 #'
 #' @inheritParams as_tibble.Spat
+#' @inheritParams tibble::print.tbl_df
+#' @param width Width of output: defaults to the setting of the width option
+#'   (if finite) or the width of the console. See [dplyr::glimpse()].
 #' @param ... Arguments passed on to [`as_tibble()`][as_tibble.Spat] Spat
 #'   methods.
-#' @param width  Width of output: defaults to the setting of the
-#'   `width` if finite (see [dplyr::glimpse()]) or the width of the console.
-#' @section terra equivalent:
+#' @param n Maximum number of rows to show.
+#'
+#' @section \CRANpkg{terra} equivalent:
 #'
 #' `print()`
+#'
 #'
 #' @section Methods:
 #'
 #' Implementation of the **generic** [dplyr::glimpse()] function for
-#' Spat*. objects.
+#' `Spat*`. objects.
 #'
 #'
 #' @examples
@@ -42,7 +46,7 @@
 #' # SpatVector
 #' v <- vect(system.file("extdata/cyl.gpkg", package = "tidyterra"))
 #'
-#' v %>% glimpse(width = 200)
+#' v %>% glimpse(n = 2)
 #'
 #' # Use on a pipeline
 #' v %>%
@@ -62,183 +66,68 @@
 #'   mutate(b = elevation_m / 100) %>%
 #'   # With options
 #'   glimpse(xy = TRUE)
-glimpse.SpatRaster <- function(x, width = NULL, ...) {
-  # Dimensions
-  cli::cat_line("Rows: ", format(terra::nrow(x),
-    big.mark = ",",
-    decimal.mark = "."
-  ))
-  cli::cat_line("Columns: ", format(terra::ncol(x),
-    big.mark = ",",
-    decimal.mark = "."
-  ))
-  cli::cat_line("Layers: ", format(terra::nlyr(x),
-    big.mark = ",",
-    decimal.mark = "."
-  ))
-  cli::cat_line("Cells: ", format(terra::ncell(x),
-    big.mark = ",",
-    decimal.mark = "."
-  ))
+glimpse.SpatRaster <- function(x, width = NULL, ..., n = NULL) {
+  # Class
+  nr <- format(terra::nrow(x), big.mark = ",", decimal.mark = ".")
+  nc <- format(terra::ncol(x), big.mark = ",", decimal.mark = ".")
+  nl <- format(terra::nlyr(x), big.mark = ",", decimal.mark = ".")
+  lay <- ifelse(terra::nlyr(x) == 1, " layer", " layers")
+  ncll <- format(terra::ncell(x), big.mark = ",", decimal.mark = ".")
 
-  if (isTRUE(sf::st_is_longlat(pull_crs(x)))) {
-    rs <- lapply(terra::res(x), decimal_to_degrees, type = "null")
-    rs <- paste0(unlist(rs), collapse = " , ")
-  } else {
-    rs <- paste(
-      format(terra::res(x),
-        big.mark = ",",
-        decimal.mark = "."
-      ),
-      collapse = " , "
-    )
-  }
-  cli::cat_line("Resolution (x , y): ", rs)
+  tterra_header(
+    "A SpatRaster ", nr, " x ", nc, " x ", nl, lay, " (",
+    ncll, " cells)"
+  )
 
+  # Resolution
+  tterra_header_string_res(x)
   # CRS
-  crsnamed <- get_named_crs(x)
-
-  if (!is.na(crsnamed)) {
-    pulled_crs <- pull_crs(x)
-    if (sf::st_is_longlat(pulled_crs)) {
-      cli::cat_line("Geodetic CRS: ", crsnamed)
-    } else {
-      cli::cat_line("Projected CRS: ", crsnamed)
-      unts <- try(sf::st_crs(pulled_crs)$units, silent = TRUE)
-      if (!inherits(unts, "try-error") && !is.null(unts) && !is.na(unts)) {
-        cli::cat_line("CRS projection units: ", unts)
-      }
-    }
-  } else {
-    cli::cat_line("CRS: Not Defined / Empty")
-  }
-
+  tterra_header_string_crs(x)
   # Extent
-  ext <- as.vector(terra::ext(x))
-  is_lonlat <- sf::st_is_longlat(pull_crs(x))
+  tterra_header_string_ext(x)
+  # RGB
+  tterra_header_string_rgb(x)
+  # Coltab
+  tterra_header_string_ctab(x)
 
-  if (isTRUE(is_lonlat)) {
-    lons <- lapply(ext[c("xmin", "xmax")], decimal_to_degrees, type = "lon")
 
-    lats <- lapply(ext[c("ymin", "ymax")], decimal_to_degrees, type = "lat")
 
-    ext_fmt <- unlist(c(lons, lats))
+  # Body
+  cli::cat_line() # Empty line
+
+  if (!terra::hasValues(x)) {
+    cli::cat_line("SpatRaster with no values")
   } else {
-    ext_fmt <- format(ext,
-      big.mark = ",",
-      decimal.mark = ".", justify = "right"
-    )
+    tterra_body(x, ..., width = width, n = n)
   }
 
-  xfmt <- paste(ext_fmt[c("xmin", "xmax")],
-    collapse = " - "
-  )
-  yfmt <- paste(ext_fmt[c("ymin", "ymax")],
-    collapse = " - "
-  )
-
-  extnamed <- paste0("[", xfmt, "] , [", yfmt, "]")
-  cli::cat_line("Extent (x , y) : ", extnamed)
-
-  # Check RGB
-  rgb_info <- terra::RGB(x)
-  if (!is.null(rgb_info) && length(rgb_info) >= 1) {
-    title <- paste0("Raster with ", length(rgb_info), " RGB channels: ")
-    ch_name <- names(x)[rgb_info]
-    names(ch_name) <- c(
-      "Red", "Green",
-      "Blue", "Alpha"
-    )[seq_len(length(rgb_info))]
-    ch_end <- paste0(ch_name, " (", names(ch_name), ")", collapse = ", ")
-
-    cli::cat_line(title, ch_end)
-  }
-
-  # Check coltab
-  coltab_info <- terra::has.colors(x)
-  if (any(coltab_info)) {
-    lcol <- length(coltab_info[coltab_info == TRUE])
-    title <- paste0("Raster with ", lcol, " color table in: ")
-    if (lcol > 1) title <- gsub("table", "tables", title)
-
-
-    ch_name <- names(x)[coltab_info == TRUE]
-    ch_end <- paste0(ch_name, collapse = ", ")
-
-    cli::cat_line(title, ch_end)
-  }
-
-
-  # Regular data frame (with options if provided)
-  capt <- utils::capture.output(dplyr::glimpse(as_tibble(x, ...),
-    width = width
-  ))
-
-  l <- ifelse(terra::nlyr(x) > 1, "Layers:", "Layer:")
-
-  cli::cat_line(l)
-  cli::cat_line(capt[-c(1, 2)])
 
   return(invisible(x))
 }
 
 #' @rdname glimpse.Spat
 #' @export
-glimpse.SpatVector <- function(x, width = NULL, ...) {
-  # Geometry type
-  cli::cat_line("Geometry type: ", tools::toTitleCase(terra::geomtype(x)))
+glimpse.SpatVector <- function(x, width = NULL, ..., n = NULL) {
+  # Class
+  nr <- format(terra::nrow(x), big.mark = ",", decimal.mark = ".")
+  nc <- format(terra::ncol(x), big.mark = ",", decimal.mark = ".")
 
-  # CRS
-  crsnamed <- get_named_crs(x)
+  tterra_header("A SpatVector ", nr, " x ", nc)
+  # Geom type
+  tterra_header("Geometry type: ", tools::toTitleCase(terra::geomtype(x)))
+  # CRS info
+  tterra_header_string_crs(x)
+  # Extent info
+  tterra_header_string_ext(x)
 
-  if (!is.na(crsnamed)) {
-    pulled_crs <- pull_crs(x)
-    if (sf::st_is_longlat(pulled_crs)) {
-      cli::cat_line("Geodetic CRS: ", crsnamed)
-    } else {
-      cli::cat_line("Projected CRS: ", crsnamed)
-      unts <- try(sf::st_crs(pulled_crs)$units, silent = TRUE)
-      if (!inherits(unts, "try-error") && !is.null(unts) && !is.na(unts)) {
-        cli::cat_line("CRS projection units: ", unts)
-      }
-    }
-  } else {
-    cli::cat_line("CRS: Not Defined / Empty")
-  }
+  # Body
+  cli::cat_line() # Empty line
 
-  # Extent
-  ext <- as.vector(terra::ext(x))
-  is_lonlat <- sf::st_is_longlat(pull_crs(x))
-
-  if (isTRUE(is_lonlat)) {
-    lons <- lapply(ext[c("xmin", "xmax")], decimal_to_degrees, type = "lon")
-
-    lats <- lapply(ext[c("ymin", "ymax")], decimal_to_degrees, type = "lat")
-
-    ext_fmt <- unlist(c(lons, lats))
-  } else {
-    ext_fmt <- format(ext,
-      big.mark = ",",
-      decimal.mark = ".", justify = "right"
-    )
-  }
-
-  xfmt <- paste(ext_fmt[c("xmin", "xmax")],
-    collapse = " - "
-  )
-  yfmt <- paste(ext_fmt[c("ymin", "ymax")],
-    collapse = " - "
-  )
-
-  extnamed <- paste0("[", xfmt, "] , [", yfmt, "]")
-
-
-  cli::cat_line("Extent (x , y) : ", extnamed)
   if (ncol(x) == 0) {
     cli::cat_line("SpatVector with no attributes (only geometries)")
   } else {
-    # Regular data frame (with options if provided)
-    dplyr::glimpse(as_tibble(x, ...), width = width)
+    # Manipulate tibble format (with options if provided)
+    tterra_body(x, ..., width = width, n = n)
   }
   return(invisible(x))
 }
@@ -248,6 +137,7 @@ glimpse.SpatVector <- function(x, width = NULL, ...) {
 dplyr::glimpse
 
 
+# Helpers ----
 
 get_named_crs <- function(x) {
   # Based in terra:::.name_or_proj4()
@@ -297,10 +187,10 @@ decimal_to_degrees <- function(x, type = c("lon", "lat", "null")) {
   type <- match.arg(type)
   coordinit <- x
   x <- abs(x)
-  D <- as.integer(x)
-  m <- (x - D) * 60
-  M <- as.integer(m)
-  S <- round((m - M) * 60, 4)
+  x_int <- as.integer(x)
+  m <- (x - x_int) * 60
+  m_int <- as.integer(m)
+  s <- round((m - m_int) * 60, 2)
 
   if (type == "lon") {
     if (coordinit > 0) {
@@ -319,15 +209,172 @@ decimal_to_degrees <- function(x, type = c("lon", "lat", "null")) {
   }
 
   if (type %in% c("lon", "lat")) {
-    label <- paste0(D, "\u00b0 ", M, "' ", S, '\" ', lab)
+    label <- paste0(x_int, "\u00b0 ", m_int, "' ", s, '\" ', lab)
   } else {
     label <- paste0(
-      c(D, M, S),
+      c(x_int, m_int, s),
       c("\u00b0", "'", '\"')
     )
-    label <- label[c(D, M, S) != 0]
+    label <- label[c(x_int, m_int, s) != 0]
     label <- paste0(label, collapse = " ")
   }
 
   return(label)
+}
+
+
+# Main style
+tterra_head_style <- cli::make_ansi_style(grey(0.6), grey = TRUE)
+
+tterra_header <- function(...) {
+  fmted <- paste("# ", paste0(..., collapse = " "))
+
+  cli::cat_line(tterra_head_style(fmted))
+}
+
+# For CRS
+tterra_header_string_crs <- function(x) {
+  crsnamed <- get_named_crs(x)
+  if (is.na(crsnamed)) {
+    tterra_header("CRS: Not Defined / Empty")
+    return(invisible(NULL))
+  }
+  pulled_crs <- pull_crs(x)
+  if (sf::st_is_longlat(pulled_crs)) {
+    tterra_header("Geodetic CRS: ", crsnamed)
+    return(invisible(NULL))
+  }
+
+  unts <- try(sf::st_crs(pulled_crs)$units, silent = TRUE)
+  # Inform of units
+  tterra_header("Projected CRS: ", crsnamed)
+
+  if (inherits(unts, "character")) {
+    unitsdb <- unitsdb
+
+    longname <- as.vector(unitsdb[unitsdb$abb == unts, ]$name)
+
+    tterra_header("CRS projection units: ", longname, " <", unts, ">")
+  }
+}
+
+# For extent
+tterra_header_string_ext <- function(x) {
+  ext <- as.vector(terra::ext(x))
+  is_lonlat <- sf::st_is_longlat(pull_crs(x))
+
+  if (isTRUE(is_lonlat)) {
+    lons <- lapply(ext[c("xmin", "xmax")], decimal_to_degrees, type = "lon")
+
+    lats <- lapply(ext[c("ymin", "ymax")], decimal_to_degrees, type = "lat")
+
+    ext_fmt <- unlist(c(lons, lats))
+  } else {
+    ext_fmt <- format(ext,
+      big.mark = ",", decimal.mark = ".",
+      justify = "right"
+    )
+  }
+
+  xfmt <- paste(ext_fmt[c("xmin", "xmax")], collapse = " / ")
+  yfmt <- paste(ext_fmt[c("ymin", "ymax")], collapse = " / ")
+
+  extnamed <- paste0("([", xfmt, "] , [", yfmt, "])")
+
+  tterra_header("Extent (x / y) : ", extnamed)
+}
+
+tterra_header_string_res <- function(x) {
+  if (isTRUE(sf::st_is_longlat(pull_crs(x)))) {
+    rs <- lapply(terra::res(x), decimal_to_degrees, type = "null")
+    rs <- paste0(unlist(rs), collapse = " , ")
+  } else {
+    rs <- paste(format(terra::res(x), big.mark = ",", decimal.mark = "."),
+      collapse = " / "
+    )
+  }
+  tterra_header("Resolution (x / y): (", rs, ")")
+}
+
+tterra_header_string_rgb <- function(x) {
+  # Check RGB color
+  rgb_info <- terra::RGB(x)
+  if (!all(!is.null(rgb_info), length(rgb_info) > 0)) {
+    return(invisible(NULL))
+  }
+
+  # Make msg
+  ch_name <- names(x)[rgb_info]
+  nm <- c("Red", "Green", "Blue", "Alpha")[seq_len(length(rgb_info))]
+  ch_end <- paste0(ch_name, " (", nm, ")", collapse = ", ")
+  pl <- ifelse(length(rgb_info) == 1, "channel", "channels")
+
+  tterra_header(
+    "SpatRaster with ", length(rgb_info), " RGB ", pl, ": ",
+    ch_end
+  )
+}
+
+tterra_header_string_ctab <- function(x) {
+  # Check coltab
+  coltab_info <- terra::has.colors(x)
+  if (!any(coltab_info)) {
+    return(invisible(NULL))
+  }
+
+  lcol <- length(coltab_info[coltab_info == TRUE])
+  pl <- ifelse(lcol == 1, "table", "tables")
+
+  ch_name <- names(x)[coltab_info == TRUE]
+  ch_end <- paste0(ch_name, collapse = ", ")
+
+  tterra_header("SpatRaster with ", lcol, " color ", pl, " in: ", ch_end)
+}
+
+# Body from tbl
+tterra_body <- function(x, width = cli::console_width(), n = 10, ...) {
+  init_type <- class(x)
+  x <- as_tibble(x, ...)
+  if (!is.numeric(n)) n <- 10
+  n <- max(1, n)
+
+  extra_cols <- NULL
+
+  if (terra::ncol(x) <= n) {
+    col_sel <- x
+  } else {
+    col_sel <- x[, seq_len(n)]
+    extra_cols <- x[, setdiff(names(x), names(col_sel))]
+  }
+
+  # Main body via glimpse
+  capt <- utils::capture.output(dplyr::glimpse(col_sel, width = width))
+  cli::cat_line(capt[-c(1:2)])
+
+  # Make footer
+  if (!is.null(extra_cols)) {
+    extra_text <- vapply(extra_cols, function(x) {
+      if (requireNamespace("vctrs", quietly = TRUE)) {
+        aa <- paste0(vctrs::vec_ptype_abbr(x), collapse = "/")
+      } else {
+        # nocov start
+        aa <- paste0(class(x), collapse = "/")
+        # nocov end
+      }
+
+      paste0("<", aa, ">")
+    }, character(1))
+
+    extra_text <- paste(names(extra_text), extra_text, collapse = ", ")
+
+    # Full message
+    nms <- ifelse(ncol(extra_cols) == 1, "variable", "variables")
+    if (init_type == "SpatRaster") nms <- gsub("variable", "layer", nms)
+    full <- paste(
+      "# ", cli::symbol$info,
+      format(ncol(extra_cols), big.mark = ".", decimal.mark = ","),
+      "more", nms, ":", extra_text
+    )
+    cli::cli_text(tterra_head_style(full))
+  }
 }