+++ R/tidyterra-package.R
@@ -3,7 +3,7 @@
 
 #' @importFrom ggplot2 aes after_stat
 #' @importFrom rlang .data
-#' @importFrom grDevices terrain.colors rgb col2rgb
+#' @importFrom grDevices terrain.colors rgb col2rgb grey
 #' @importFrom tibble as_tibble
 #' @importFrom cli qty
 NULL